refactor(aggregator): move `GcpBackendUploaderLegacy::get_location` to `CloudRemotePath`
* Renamed to `to_gcloud_storage_location`
* This allow easy reuse of this code with the new backend
