+++ mithril-aggregator/src/file_uploaders/gcp_uploader/interface.rs
@@ -20,6 +20,19 @@ impl CloudRemotePath {
     pub fn join(&self, file_path: &str) -> Self {
         Self(self.0.join(file_path))
     }
+
+    /// Transform this cloud remote path to file uri hosted in Google Cloud storage
+    pub fn to_gcloud_storage_location(&self, bucket: &str, use_cdn_domain: bool) -> FileUri {
+        let mut uri = vec![];
+        if !use_cdn_domain {
+            uri.push("storage.googleapis.com");
+        }
+        uri.push(bucket);
+        let file_path = self.to_string();
+        uri.push(&file_path);
+
+        FileUri(format!("https://{}", uri.join("/")))
+    }
 }
 
 impl Display for CloudRemotePath {
@@ -50,4 +63,37 @@ pub trait CloudBackendUploader: Send + Sync {
 
     /// Make a file public in the cloud backend
     async fn make_file_public(&self, remote_file_path: &CloudRemotePath) -> StdResult<()>;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn get_location_not_using_cdn_domain_return_google_api_uri() {
+        let use_cdn_domain = false;
+        let bucket = "cdn.mithril.network".to_string();
+        let remote_file_path = CloudRemotePath::new("remote_folder").join("snapshot.xxx.tar.gz");
+
+        let location = remote_file_path.to_gcloud_storage_location(&bucket, use_cdn_domain);
+
+        assert_eq!(
+            FileUri("https://storage.googleapis.com/cdn.mithril.network/remote_folder/snapshot.xxx.tar.gz".to_string()),
+            location
+        );
+    }
+
+    #[tokio::test]
+    async fn get_location_using_cdn_domain_return_cdn_in_uri() {
+        let use_cdn_domain = true;
+        let bucket = "cdn.mithril.network".to_string();
+        let remote_file_path = CloudRemotePath::new("remote_folder").join("snapshot.xxx.tar.gz");
+
+        let location = remote_file_path.to_gcloud_storage_location(&bucket, use_cdn_domain);
+
+        assert_eq!(
+            FileUri("https://cdn.mithril.network/remote_folder/snapshot.xxx.tar.gz".to_string()),
+            location
+        );
+    }
 }
